modified parser to use CensusGroup[] instead of CensusData and Doubles instead of floats

+++ src/cse332/types/CensusData.java
@@ -13,7 +13,7 @@ public class CensusData {
 		data_size = 0;
 	}
 	
-	public void add(int population, float latitude, float longitude) {
+	public void add(int population, double latitude, double longitude) {
 		if(data_size == data.length) { // resize
 			CensusGroup[] new_data = new CensusGroup[data.length*2];
 			for(int i=0; i < data.length; ++i)
